codigos da aula 5

+++ src/main/java/br/com/alura/carteira/service/TransacaoService.java
@@ -6,6 +6,7 @@ import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -30,18 +31,23 @@ public class TransacaoService {
 	@Autowired
 	private ModelMapper modelMapper;
 
-	public Page<TransacaoDto> listar(Pageable paginacao) {
+	public Page<TransacaoDto> listar(Pageable paginacao, Usuario logado) {
 		return repository
-				.findAll(paginacao)
+				.findAllByUsuario(logado, paginacao)
 				.map(t -> modelMapper.map(t, TransacaoDto.class));
 	}
 
 	@Transactional
-	public TransacaoDto cadastrar(TransacaoFormDto dto) {
+	public TransacaoDto cadastrar(TransacaoFormDto dto, Usuario logado) {
 		Long idUsuario = dto.getUsuarioId();
 		
 		try {
 			Usuario usuario = usuarioRepository.getById(idUsuario);
+			
+			if (!logado.equals(usuario)) {
+				lancarErro403();
+			}
+			
 			Transacao transacao = modelMapper.map(dto, Transacao.class);
 			transacao.setId(null);
 			transacao.setUsuario(usuario);
@@ -55,24 +61,44 @@ public class TransacaoService {
 	}
 
 	@Transactional
-	public TransacaoDto atualizar(AtualizacaoTransacaoFormDto dto) {
+	public TransacaoDto atualizar(AtualizacaoTransacaoFormDto dto, Usuario logado) {
 		Transacao transacao = repository.getById(dto.getId());
+		
+		if (!transacao.pertenceAo(logado)) {
+			lancarErro403();
+		}
+		
 		transacao.atualizarInformacoes(dto.getTicker(), dto.getData(), dto.getPreco(), dto.getQuantidade(), dto.getTipo());
 		
 		return modelMapper.map(transacao, TransacaoDto.class);
 	}
 
 	@Transactional
-	public void remover(Long id) {
-		repository.deleteById(id);
-	}
-
-	public TransacaoDetalhadaDto detalhar(Long id) {
+	public void remover(Long id, Usuario logado) {
 		Transacao transacao = repository
 				.findById(id)
 				.orElseThrow(() -> new EntityNotFoundException());
 		
+		if (transacao.pertenceAo(logado)) {
+			repository.deleteById(id);
+		} else {
+			lancarErro403();
+		}
+	}
+
+	public TransacaoDetalhadaDto detalhar(Long id, Usuario logado) {
+		Transacao transacao = repository
+				.findById(id)
+				.orElseThrow(() -> new EntityNotFoundException());
+		
+		if (!transacao.pertenceAo(logado)) {
+			lancarErro403();
+		}
 		return modelMapper.map(transacao, TransacaoDetalhadaDto.class);
+	}
+
+	private void lancarErro403() {
+		throw new AccessDeniedException("Acesso negado!");
 	}
 
 }
